Revert "refactor: prover inputs (#646)"
This reverts commit 80182bfdea1ad09ea885f1caff41b81557305ed6.

+++ gnark-prover/prover/marshal.go
@@ -97,28 +97,28 @@ func (ps *ProvingSystem) WriteTo(w io.Writer) (int64, error) {
 	var totalWritten int64 = 0
 	var intBuf [4]byte
 
-	binary.BigEndian.PutUint32(intBuf[:], uint32(ps.InclusionTreeDepth))
+	binary.BigEndian.PutUint32(intBuf[:], ps.InclusionTreeDepth)
 	written, err := w.Write(intBuf[:])
 	totalWritten += int64(written)
 	if err != nil {
 		return totalWritten, err
 	}
 
-	binary.BigEndian.PutUint32(intBuf[:], uint32(ps.InclusionNumberOfUtxos))
+	binary.BigEndian.PutUint32(intBuf[:], ps.InclusionNumberOfUtxos)
 	written, err = w.Write(intBuf[:])
 	totalWritten += int64(written)
 	if err != nil {
 		return totalWritten, err
 	}
 
-	binary.BigEndian.PutUint32(intBuf[:], uint32(ps.NonInclusionTreeDepth))
+	binary.BigEndian.PutUint32(intBuf[:], ps.NonInclusionTreeDepth)
 	written, err = w.Write(intBuf[:])
 	totalWritten += int64(written)
 	if err != nil {
 		return totalWritten, err
 	}
 
-	binary.BigEndian.PutUint32(intBuf[:], uint32(ps.NonInclusionNumberOfUtxos))
+	binary.BigEndian.PutUint32(intBuf[:], ps.NonInclusionNumberOfUtxos)
 	written, err = w.Write(intBuf[:])
 	totalWritten += int64(written)
 	if err != nil {
